Move diff, fillna, shift
