+++ python/cudf/cudf/tests/test_groupby.py
@@ -26,7 +26,6 @@ from cudf.testing import assert_eq
 from cudf.testing._utils import (
     expect_warning_if,
 )
-from cudf.testing.dataset_generator import rand_dataframe
 
 _now = np.datetime64("now")
 _tomorrow = _now + np.timedelta64(1, "D")
@@ -837,555 +836,6 @@ def test_groupby_apply_no_keys(pdf):
         gdf.groupby([]).apply(lambda x: x.max()),
         check_index_type=False,  # Int64 v/s Float64
         **kwargs,
-    )
-
-
-@pytest.mark.parametrize("method", ["average", "min", "max", "first", "dense"])
-@pytest.mark.parametrize("ascending", [True, False])
-@pytest.mark.parametrize("na_option", ["keep", "top", "bottom"])
-@pytest.mark.parametrize("pct", [False, True])
-def test_groupby_2keys_rank(method, ascending, na_option, pct):
-    nelem = 20
-    t = rand_dataframe(
-        dtypes_meta=[
-            {"dtype": "int64", "null_frequency": 0, "cardinality": 10},
-            {"dtype": "int64", "null_frequency": 0, "cardinality": 10},
-            {"dtype": "int64", "null_frequency": 0.4, "cardinality": 10},
-        ],
-        rows=nelem,
-        use_threads=False,
-        seed=0,
-    )
-    pdf = t.to_pandas()
-    pdf.columns = ["x", "y", "z"]
-    gdf = cudf.from_pandas(pdf)
-    expect_df = pdf.groupby(["x", "y"], sort=True).rank(
-        method=method, ascending=ascending, na_option=na_option, pct=pct
-    )
-    got_df = gdf.groupby(["x", "y"], sort=True).rank(
-        method=method, ascending=ascending, na_option=na_option, pct=pct
-    )
-
-    assert_groupby_results_equal(got_df, expect_df, check_dtype=False)
-
-
-def test_groupby_rank_fails():
-    gdf = cudf.DataFrame(
-        {"x": [1, 2, 3, 4], "y": [1, 2, 3, 4], "z": [1, 2, 3, 4]}
-    )
-    with pytest.raises(NotImplementedError):
-        gdf.groupby(["x", "y"]).rank(method="min", axis=1)
-    gdf = cudf.DataFrame(
-        {
-            "a": [1, 1, 1, 2, 2, 2],
-            "b": [[1, 2], [3, None, 5], None, [], [7, 8], [9]],
-        }
-    )
-    with pytest.raises(NotImplementedError):
-        gdf.groupby(["a"]).rank(method="min", axis=1)
-
-
-@pytest.mark.parametrize(
-    "with_nan", [False, True], ids=["just-NA", "also-NaN"]
-)
-@pytest.mark.parametrize("dropna", [False, True], ids=["keepna", "dropna"])
-@pytest.mark.parametrize(
-    "duplicate_index", [False, True], ids=["rangeindex", "dupindex"]
-)
-def test_groupby_scan_null_keys(with_nan, dropna, duplicate_index):
-    key_col = [None, 1, 2, None, 3, None, 3, 1, None, 1]
-    if with_nan:
-        df = pd.DataFrame(
-            {"key": pd.Series(key_col, dtype="float32"), "value": range(10)}
-        )
-    else:
-        df = pd.DataFrame(
-            {"key": pd.Series(key_col, dtype="Int32"), "value": range(10)}
-        )
-
-    if duplicate_index:
-        # Non-default index with duplicates
-        df.index = [1, 2, 3, 1, 3, 2, 4, 1, 6, 10]
-
-    cdf = cudf.from_pandas(df)
-
-    expect = df.groupby("key", dropna=dropna).cumsum()
-    got = cdf.groupby("key", dropna=dropna).cumsum()
-    assert_eq(expect, got)
-
-
-def test_groupby_mix_agg_scan():
-    err_msg = "Cannot perform both aggregation and scan in one operation"
-    func = ["cumsum", "sum"]
-    gb = make_frame(DataFrame, nelem=10).groupby(["x", "y"], sort=True)
-
-    gb.agg(func[0])
-    gb.agg(func[1])
-    gb.agg(func[1:])
-    with pytest.raises(NotImplementedError, match=err_msg):
-        gb.agg(func)
-
-
-@pytest.mark.parametrize("shift_perc", [0.5, 1.0, 1.5])
-@pytest.mark.parametrize("direction", [1, -1])
-@pytest.mark.parametrize("fill_value", [None, np.nan, 42])
-def test_groupby_shift_row(shift_perc, direction, fill_value):
-    nelem = 20
-    pdf = make_frame(pd.DataFrame, nelem=nelem, extra_vals=["val2"])
-    gdf = cudf.from_pandas(pdf)
-    n_shift = int(nelem * shift_perc) * direction
-
-    expected = pdf.groupby(["x", "y"]).shift(
-        periods=n_shift, fill_value=fill_value
-    )
-    got = gdf.groupby(["x", "y"]).shift(periods=n_shift, fill_value=fill_value)
-
-    assert_groupby_results_equal(
-        expected[["val", "val2"]], got[["val", "val2"]]
-    )
-
-
-@pytest.mark.parametrize("shift_perc", [0.5, 1.0, 1.5])
-@pytest.mark.parametrize("direction", [1, -1])
-@pytest.mark.parametrize(
-    "fill_value",
-    [
-        None,
-        pytest.param(
-            0,
-            marks=pytest.mark.xfail(
-                reason="https://github.com/rapidsai/cudf/issues/10608"
-            ),
-        ),
-        pytest.param(
-            42,
-            marks=pytest.mark.xfail(
-                reason="https://github.com/rapidsai/cudf/issues/10608"
-            ),
-        ),
-    ],
-)
-def test_groupby_shift_row_mixed_numerics(shift_perc, direction, fill_value):
-    nelem = 20
-    t = rand_dataframe(
-        dtypes_meta=[
-            {"dtype": "int64", "null_frequency": 0, "cardinality": 10},
-            {"dtype": "int64", "null_frequency": 0.4, "cardinality": 10},
-            {"dtype": "float32", "null_frequency": 0.4, "cardinality": 10},
-            {
-                "dtype": "datetime64[ns]",
-                "null_frequency": 0.4,
-                "cardinality": 10,
-            },
-            {
-                "dtype": "timedelta64[ns]",
-                "null_frequency": 0.4,
-                "cardinality": 10,
-            },
-        ],
-        rows=nelem,
-        use_threads=False,
-        seed=0,
-    )
-    pdf = t.to_pandas()
-    gdf = cudf.from_pandas(pdf)
-    n_shift = int(nelem * shift_perc) * direction
-
-    expected = pdf.groupby(["0"]).shift(periods=n_shift, fill_value=fill_value)
-    got = gdf.groupby(["0"]).shift(periods=n_shift, fill_value=fill_value)
-
-    assert_groupby_results_equal(
-        expected[["1", "2", "3", "4"]], got[["1", "2", "3", "4"]]
-    )
-
-
-# TODO: Shifting list columns is currently unsupported because we cannot
-# construct a null list scalar in python. Support once it is added.
-@pytest.mark.parametrize("shift_perc", [0.5, 1.0, 1.5])
-@pytest.mark.parametrize("direction", [1, -1])
-def test_groupby_shift_row_mixed(shift_perc, direction):
-    nelem = 20
-    t = rand_dataframe(
-        dtypes_meta=[
-            {"dtype": "int64", "null_frequency": 0, "cardinality": 10},
-            {"dtype": "int64", "null_frequency": 0.4, "cardinality": 10},
-            {"dtype": "str", "null_frequency": 0.4, "cardinality": 10},
-            {
-                "dtype": "datetime64[ns]",
-                "null_frequency": 0.4,
-                "cardinality": 10,
-            },
-            {
-                "dtype": "timedelta64[ns]",
-                "null_frequency": 0.4,
-                "cardinality": 10,
-            },
-        ],
-        rows=nelem,
-        use_threads=False,
-        seed=0,
-    )
-    pdf = t.to_pandas()
-    gdf = cudf.from_pandas(pdf)
-    n_shift = int(nelem * shift_perc) * direction
-
-    expected = pdf.groupby(["0"]).shift(periods=n_shift)
-    got = gdf.groupby(["0"]).shift(periods=n_shift)
-
-    assert_groupby_results_equal(
-        expected[["1", "2", "3", "4"]], got[["1", "2", "3", "4"]]
-    )
-
-
-@pytest.mark.parametrize("shift_perc", [0.5, 1.0, 1.5])
-@pytest.mark.parametrize("direction", [1, -1])
-@pytest.mark.parametrize(
-    "fill_value",
-    [
-        [
-            42,
-            "fill",
-            np.datetime64(123, "ns"),
-            np.timedelta64(456, "ns"),
-        ]
-    ],
-)
-def test_groupby_shift_row_mixed_fill(shift_perc, direction, fill_value):
-    nelem = 20
-    t = rand_dataframe(
-        dtypes_meta=[
-            {"dtype": "int64", "null_frequency": 0, "cardinality": 10},
-            {"dtype": "int64", "null_frequency": 0.4, "cardinality": 10},
-            {"dtype": "str", "null_frequency": 0.4, "cardinality": 10},
-            {
-                "dtype": "datetime64[ns]",
-                "null_frequency": 0.4,
-                "cardinality": 10,
-            },
-            {
-                "dtype": "timedelta64[ns]",
-                "null_frequency": 0.4,
-                "cardinality": 10,
-            },
-        ],
-        rows=nelem,
-        use_threads=False,
-        seed=0,
-    )
-    pdf = t.to_pandas()
-    gdf = cudf.from_pandas(pdf)
-    n_shift = int(nelem * shift_perc) * direction
-
-    # Pandas does not support specifying different fill_value by column, so we
-    # simulate it column by column
-    expected = pdf.copy()
-    for col, single_fill in zip(pdf.iloc[:, 1:], fill_value, strict=True):
-        expected[col] = (
-            pdf[col]
-            .groupby(pdf["0"])
-            .shift(periods=n_shift, fill_value=single_fill)
-        )
-
-    got = gdf.groupby(["0"]).shift(periods=n_shift, fill_value=fill_value)
-
-    assert_groupby_results_equal(
-        expected[["1", "2", "3", "4"]], got[["1", "2", "3", "4"]]
-    )
-
-
-@pytest.mark.parametrize("fill_value", [None, 0, 42])
-def test_groupby_shift_row_zero_shift(fill_value):
-    nelem = 20
-    t = rand_dataframe(
-        dtypes_meta=[
-            {"dtype": "int64", "null_frequency": 0, "cardinality": 10},
-            {"dtype": "int64", "null_frequency": 0.4, "cardinality": 10},
-            {"dtype": "float32", "null_frequency": 0.4, "cardinality": 10},
-            {
-                "dtype": "datetime64[ns]",
-                "null_frequency": 0.4,
-                "cardinality": 10,
-            },
-            {
-                "dtype": "timedelta64[ns]",
-                "null_frequency": 0.4,
-                "cardinality": 10,
-            },
-        ],
-        rows=nelem,
-        use_threads=False,
-        seed=0,
-    )
-    gdf = cudf.from_pandas(t.to_pandas())
-
-    expected = gdf
-    got = gdf.groupby(["0"]).shift(periods=0, fill_value=fill_value)
-
-    assert_groupby_results_equal(
-        expected[["1", "2", "3", "4"]], got[["1", "2", "3", "4"]]
-    )
-
-
-@pytest.mark.parametrize("shift_perc", [0.5, 1.0, 1.5])
-@pytest.mark.parametrize("direction", [1, -1])
-def test_groupby_diff_row(shift_perc, direction):
-    nelem = 20
-    pdf = make_frame(pd.DataFrame, nelem=nelem, extra_vals=["val2"])
-    gdf = cudf.from_pandas(pdf)
-    n_shift = int(nelem * shift_perc) * direction
-
-    expected = pdf.groupby(["x", "y"]).diff(periods=n_shift)
-    got = gdf.groupby(["x", "y"]).diff(periods=n_shift)
-
-    assert_groupby_results_equal(
-        expected[["val", "val2"]], got[["val", "val2"]]
-    )
-
-
-@pytest.mark.parametrize("shift_perc", [0.5, 1.0, 1.5])
-@pytest.mark.parametrize("direction", [1, -1])
-def test_groupby_diff_row_mixed_numerics(shift_perc, direction):
-    nelem = 20
-    t = rand_dataframe(
-        dtypes_meta=[
-            {"dtype": "int64", "null_frequency": 0, "cardinality": 10},
-            {"dtype": "int64", "null_frequency": 0.4, "cardinality": 10},
-            {"dtype": "float32", "null_frequency": 0.4, "cardinality": 10},
-            {"dtype": "decimal64", "null_frequency": 0.4, "cardinality": 10},
-            {
-                "dtype": "datetime64[ns]",
-                "null_frequency": 0.4,
-                "cardinality": 10,
-            },
-            {
-                "dtype": "timedelta64[ns]",
-                "null_frequency": 0.4,
-                "cardinality": 10,
-            },
-        ],
-        rows=nelem,
-        use_threads=False,
-        seed=0,
-    )
-    pdf = t.to_pandas()
-    gdf = cudf.from_pandas(pdf)
-    n_shift = int(nelem * shift_perc) * direction
-
-    expected = pdf.groupby(["0"]).diff(periods=n_shift)
-    got = gdf.groupby(["0"]).diff(periods=n_shift)
-
-    assert_groupby_results_equal(
-        expected[["1", "2", "3", "4", "5"]], got[["1", "2", "3", "4", "5"]]
-    )
-
-
-def test_groupby_diff_row_zero_shift():
-    nelem = 20
-    t = rand_dataframe(
-        dtypes_meta=[
-            {"dtype": "int64", "null_frequency": 0, "cardinality": 10},
-            {"dtype": "int64", "null_frequency": 0.4, "cardinality": 10},
-            {"dtype": "float32", "null_frequency": 0.4, "cardinality": 10},
-            {
-                "dtype": "datetime64[ns]",
-                "null_frequency": 0.4,
-                "cardinality": 10,
-            },
-            {
-                "dtype": "timedelta64[ns]",
-                "null_frequency": 0.4,
-                "cardinality": 10,
-            },
-        ],
-        rows=nelem,
-        use_threads=False,
-        seed=0,
-    )
-    gdf = cudf.from_pandas(t.to_pandas())
-
-    expected = gdf
-    got = gdf.groupby(["0"]).shift(periods=0)
-
-    assert_groupby_results_equal(
-        expected[["1", "2", "3", "4"]], got[["1", "2", "3", "4"]]
-    )
-
-
-@pytest.mark.skipif(
-    PANDAS_VERSION < PANDAS_CURRENT_SUPPORTED_VERSION,
-    reason="warning not present in older pandas versions",
-)
-def test_groupby_fillna_multi_value():
-    nelem = 20
-    t = rand_dataframe(
-        dtypes_meta=[
-            {"dtype": "int64", "null_frequency": 0, "cardinality": 10},
-            {"dtype": "int64", "null_frequency": 0.4, "cardinality": 10},
-            {"dtype": "float32", "null_frequency": 0.4, "cardinality": 10},
-            {
-                "dtype": "datetime64[ms]",
-                "null_frequency": 0.4,
-                "cardinality": 10,
-            },
-            {
-                "dtype": "timedelta64[ns]",
-                "null_frequency": 0.4,
-                "cardinality": 10,
-            },
-            {"dtype": "decimal64", "null_frequency": 0.4, "cardinality": 10},
-            {"dtype": "str", "null_frequency": 0.4, "cardinality": 10},
-        ],
-        rows=nelem,
-        use_threads=False,
-        seed=0,
-    )
-    key_col = "0"
-    value_cols = ["1", "2", "3", "4", "5", "6"]
-    pdf = t.to_pandas()
-    gdf = cudf.from_pandas(pdf)
-
-    # fill the dataframe with the first non-null item in the column
-    fill_values = {
-        name: pdf[name].loc[pdf[name].first_valid_index()]
-        for name in value_cols
-    }
-    # cudf can't fillna with a pandas.Timedelta type
-    fill_values["4"] = fill_values["4"].to_numpy()
-    with pytest.warns(FutureWarning):
-        expect = pdf.groupby(key_col).fillna(value=fill_values)
-    with pytest.warns(FutureWarning):
-        got = gdf.groupby(key_col).fillna(value=fill_values)
-
-    assert_groupby_results_equal(expect[value_cols], got[value_cols])
-
-
-# TODO: cudf.fillna does not support decimal column to column fill yet
-@pytest.mark.skipif(
-    PANDAS_VERSION < PANDAS_CURRENT_SUPPORTED_VERSION,
-    reason="warning not present in older pandas versions",
-)
-def test_groupby_fillna_multi_value_df():
-    nelem = 20
-    t = rand_dataframe(
-        dtypes_meta=[
-            {"dtype": "int64", "null_frequency": 0, "cardinality": 10},
-            {"dtype": "int64", "null_frequency": 0.4, "cardinality": 10},
-            {"dtype": "float32", "null_frequency": 0.4, "cardinality": 10},
-            {
-                "dtype": "datetime64[ms]",
-                "null_frequency": 0.4,
-                "cardinality": 10,
-            },
-            {
-                "dtype": "timedelta64[ns]",
-                "null_frequency": 0.4,
-                "cardinality": 10,
-            },
-            {"dtype": "str", "null_frequency": 0.4, "cardinality": 10},
-        ],
-        rows=nelem,
-        use_threads=False,
-        seed=0,
-    )
-    key_col = "0"
-    value_cols = ["1", "2", "3", "4", "5"]
-    pdf = t.to_pandas()
-    gdf = cudf.from_pandas(pdf)
-
-    # fill the dataframe with the first non-null item in the column
-    fill_values = {
-        name: pdf[name].loc[pdf[name].first_valid_index()]
-        for name in value_cols
-    }
-    # cudf can't fillna with a pandas.Timedelta type
-    fill_values["4"] = fill_values["4"].to_numpy()
-    fill_values = pd.DataFrame(fill_values, index=pdf.index)
-    with pytest.warns(FutureWarning):
-        expect = pdf.groupby(key_col).fillna(value=fill_values)
-
-    fill_values = cudf.from_pandas(fill_values)
-    with pytest.warns(FutureWarning):
-        got = gdf.groupby(key_col).fillna(value=fill_values)
-
-    assert_groupby_results_equal(expect[value_cols], got[value_cols])
-
-
-@pytest.mark.parametrize(
-    "by",
-    [pd.Series([1, 1, 2, 2, 3, 4]), lambda x: x % 2 == 0, pd.Grouper(level=0)],
-)
-@pytest.mark.parametrize(
-    "data", [[1, None, 2, None, 3, None], [1, 2, 3, 4, 5, 6]]
-)
-@pytest.mark.parametrize("args", [{"value": 42}, {"method": "ffill"}])
-@pytest.mark.skipif(
-    PANDAS_VERSION < PANDAS_CURRENT_SUPPORTED_VERSION,
-    reason="warning not present in older pandas versions",
-)
-def test_groupby_various_by_fillna(by, data, args):
-    ps = pd.Series(data)
-    gs = cudf.from_pandas(ps)
-
-    with pytest.warns(FutureWarning):
-        expect = ps.groupby(by).fillna(**args)
-    if isinstance(by, pd.Grouper):
-        by = cudf.Grouper(level=by.level)
-    with pytest.warns(FutureWarning):
-        got = gs.groupby(by).fillna(**args)
-
-    assert_groupby_results_equal(expect, got, check_dtype=False)
-
-
-@pytest.mark.skipif(
-    PANDAS_VERSION < PANDAS_CURRENT_SUPPORTED_VERSION,
-    reason="warning not present in older pandas versions",
-)
-@pytest.mark.parametrize("method", ["ffill", "bfill"])
-def test_groupby_fillna_method(method):
-    nelem = 20
-    t = rand_dataframe(
-        dtypes_meta=[
-            {"dtype": "int64", "null_frequency": 0, "cardinality": 10},
-            {"dtype": "int64", "null_frequency": 0.4, "cardinality": 10},
-            {"dtype": "float32", "null_frequency": 0.4, "cardinality": 10},
-            {
-                "dtype": "datetime64[ns]",
-                "null_frequency": 0.4,
-                "cardinality": 10,
-            },
-            {
-                "dtype": "timedelta64[ns]",
-                "null_frequency": 0.4,
-                "cardinality": 10,
-            },
-            {
-                "dtype": "list",
-                "null_frequency": 0.4,
-                "cardinality": 10,
-                "lists_max_length": 10,
-                "nesting_max_depth": 3,
-                "value_type": "int64",
-            },
-            {"dtype": "category", "null_frequency": 0.4, "cardinality": 10},
-            {"dtype": "decimal64", "null_frequency": 0.4, "cardinality": 10},
-            {"dtype": "str", "null_frequency": 0.4, "cardinality": 10},
-        ],
-        rows=nelem,
-        use_threads=False,
-        seed=0,
-    )
-    key_col = "0"
-    value_cols = ["1", "2", "3", "4", "5", "6", "7", "8"]
-    pdf = t.to_pandas()
-    gdf = cudf.from_pandas(pdf)
-
-    with pytest.warns(FutureWarning):
-        expect = pdf.groupby(key_col).fillna(method=method)
-    with pytest.warns(FutureWarning):
-        got = gdf.groupby(key_col).fillna(method=method)
-
-    assert_groupby_results_equal(
-        expect[value_cols], got[value_cols], sort=False
     )
